// cocktails.web/src/api/cocktailsApi/models/ingredientFilterModel.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Cocktails Api
 * Cocktails Api
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface IngredientFilterModel
 */
export interface IngredientFilterModel {
    /**
     * The filter identifier
     * @type {string}
     * @memberof IngredientFilterModel
     */
    id: string;
    /**
     * The display name of the filter
     * @type {string}
     * @memberof IngredientFilterModel
     */
    name: string;
}

/**
 * Check if a given object implements the IngredientFilterModel interface.
 */
export function instanceOfIngredientFilterModel(value: object): value is IngredientFilterModel {
    if (!('id' in value) || value['id'] === undefined) return false;
    if (!('name' in value) || value['name'] === undefined) return false;
    return true;
}

export function IngredientFilterModelFromJSON(json: any): IngredientFilterModel {
    return IngredientFilterModelFromJSONTyped(json, false);
}

export function IngredientFilterModelFromJSONTyped(json: any, ignoreDiscriminator: boolean): IngredientFilterModel {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'],
        'name': json['name'],
    };
}

export function IngredientFilterModelToJSON(json: any): IngredientFilterModel {
    return IngredientFilterModelToJSONTyped(json, false);
}

export function IngredientFilterModelToJSONTyped(value?: IngredientFilterModel | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'id': value['id'],
        'name': value['name'],
    };
}